Move and rename classes from the analyst directory
