+++ src/main/java/org/opentripplanner/common/geometry/MapTile.java
@@ -1,14 +1,14 @@
-package org.opentripplanner.analyst.request;
+package org.opentripplanner.common.geometry;
 
 import org.geotools.geometry.Envelope2D;
 
-public class TileRequest {
+public class MapTile {
 
     public final Envelope2D bbox; // includes CRS
     public final int width; 
     public final int height; 
 
-    public TileRequest(Envelope2D bbox, Integer width, Integer height) {
+    public MapTile(Envelope2D bbox, Integer width, Integer height) {
         this.bbox = bbox;
         this.width = width;
         this.height = height;
@@ -19,8 +19,8 @@ public class TileRequest {
     }
     
     public boolean equals(Object other) {
-        if (other instanceof TileRequest) {
-            TileRequest that = (TileRequest) other;
+        if (other instanceof MapTile) {
+            MapTile that = (MapTile) other;
             return this.bbox.equals(that.bbox) &&
                    this.width  == that.width   &&
                    this.height == that.height;
